requirements.txt
pyspark>=3.4.0
delta-spark>=2.4.0
jupyter>=1.0.0
notebook>=6.0.0
ipykernel>=6.0.0